requirements.txt
google-cloud-bigquery
google-auth
google-auth-oauthlib
pandas